rename : Layout 폴더안의 컴포넌트를 목적에 맞는 컴포넌트 폴더 위치로 변경
